Better setup for leaflet map
