+++ public/index.html
@@ -9,11 +9,15 @@
   <title>sunflower_client</title>
 
   <link rel="stylesheet" href="<%= BASE_URL %>css/bootstrap.min.css">
+  <!--
   <link rel="stylesheet" href="<%= BASE_URL %>css/leaflet.css">
+  -->
 
   <script src="<%= BASE_URL %>js/jquery-3.0.0.min.js"></script>
   <script src="<%= BASE_URL %>js/bootstrap.min.js"></script>
+  <!--
   <script src="<%= BASE_URL %>js/leaflet.js"></script>
+  -->
 </head>
 
 <body>
